# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IdentityVerificationLevel =
  (typeof IdentityVerificationLevel)[keyof typeof IdentityVerificationLevel];

export const IdentityVerificationLevel = {
  none: "none",
  device: "device",
  orb: "orb",
} as const;

export interface Identity {
  id: string;
  worldId: string;
  username: string;
  avatarUrl?: string;
  verificationLevel: IdentityVerificationLevel;
  isVerified: boolean;
  nullifierHash?: string;
  walletAddress: string;
  joinedAt: string;
}

export type InitiateVerificationBodyLevel =
  (typeof InitiateVerificationBodyLevel)[keyof typeof InitiateVerificationBodyLevel];

export const InitiateVerificationBodyLevel = {
  device: "device",
  orb: "orb",
} as const;

export interface InitiateVerificationBody {
  level: InitiateVerificationBodyLevel;
}

export type VerificationSessionStatus =
  (typeof VerificationSessionStatus)[keyof typeof VerificationSessionStatus];

export const VerificationSessionStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface VerificationSession {
  sessionId: string;
  status: VerificationSessionStatus;
  level: string;
  expiresAt: string;
}

export type CredentialType =
  (typeof CredentialType)[keyof typeof CredentialType];

export const CredentialType = {
  orb: "orb",
  device: "device",
  phone: "phone",
} as const;

export interface Credential {
  id: string;
  type: CredentialType;
  label: string;
  issuedAt: string;
  expiresAt?: string;
  isActive: boolean;
}

export interface Token {
  symbol: string;
  name: string;
  balance: number;
  balanceUsd: number;
  priceUsd: number;
  change24hPercent: number;
  iconUrl?: string;
  contractAddress?: string;
  chain: string;
}

export interface Wallet {
  address: string;
  totalValueUsd: number;
  change24hPercent: number;
  change24hUsd: number;
  tokens: Token[];
}

export interface SendTokensBody {
  toAddress: string;
  amount: number;
  token: string;
  note?: string;
}

export interface ReceiveInfo {
  address: string;
  qrCodeData: string;
  username: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  send: "send",
  receive: "receive",
  grant: "grant",
  swap: "swap",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  confirmed: "confirmed",
  failed: "failed",
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  status: TransactionStatus;
  amount: number;
  amountUsd: number;
  token: string;
  fromAddress?: string;
  toAddress?: string;
  toUsername?: string;
  fromUsername?: string;
  txHash?: string;
  note?: string;
  createdAt: string;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface TransactionSummary {
  totalSentUsd: number;
  totalReceivedUsd: number;
  totalGrantsUsd: number;
  transactionCount: number;
  thisMonthUsd: number;
}

export interface MiniApp {
  id: string;
  name: string;
  description: string;
  iconUrl?: string;
  category: string;
  categoryLabel: string;
  developer: string;
  rating: number;
  ratingCount: number;
  userCount: number;
  isVerified: boolean;
  isFeatured: boolean;
  tags: string[];
  url: string;
  screenshotUrls?: string[];
}

export interface Category {
  id: string;
  label: string;
  icon: string;
  appCount: number;
}

export interface LaunchResult {
  success: boolean;
  url: string;
}

export type GrantType = (typeof GrantType)[keyof typeof GrantType];

export const GrantType = {
  world_id_grant: "world_id_grant",
  referral: "referral",
  airdrop: "airdrop",
  reward: "reward",
} as const;

export type GrantStatus = (typeof GrantStatus)[keyof typeof GrantStatus];

export const GrantStatus = {
  available: "available",
  claimed: "claimed",
  expired: "expired",
} as const;

export interface Grant {
  id: string;
  type: GrantType;
  title: string;
  description: string;
  amountWld: number;
  amountUsd: number;
  status: GrantStatus;
  expiresAt?: string;
  requiresOrb: boolean;
}

export interface ClaimGrantBody {
  grantId: string;
}

export interface ClaimResult {
  success: boolean;
  amountWld: number;
  amountUsd: number;
  txHash?: string;
}

export interface GlobalStats {
  totalVerifiedHumans: number;
  totalCountries: number;
  totalMiniApps: number;
  totalWldDistributed: number;
  monthlyActiveUsers: number;
}

export type GetTransactionsParams = {
  page?: number;
  limit?: number;
  type?: GetTransactionsType;
};

export type GetTransactionsType =
  (typeof GetTransactionsType)[keyof typeof GetTransactionsType];

export const GetTransactionsType = {
  send: "send",
  receive: "receive",
  grant: "grant",
  swap: "swap",
} as const;

export type GetMiniAppsParams = {
  category?: string;
  featured?: boolean;
  search?: string;
};