# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: World App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mini-apps:
    get:
      operationId: getMiniApps
      tags: [mini-apps]
      summary: Get mini apps
      description: Returns available mini apps in the marketplace
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Mini apps list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MiniApp"
...
  /mini-apps/{id}:
    get:
      operationId: getMiniApp
      tags: [mini-apps]
      summary: Get mini app details
      description: Returns details about a specific mini app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mini app details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiniApp"
  /mini-apps/{id}/launch:
    post:
      operationId: launchMiniApp
      tags: [mini-apps]
      summary: Launch mini app
      description: Record that the user launched a mini app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Launch recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LaunchResult"
...
components:
  schemas:
...
    MiniApp:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        iconUrl:
          type: string
        category:
          type: string
        categoryLabel:
          type: string
        developer:
          type: string
        rating:
          type: number
        ratingCount:
          type: integer
        userCount:
          type: integer
        isVerified:
          type: boolean
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        url:
          type: string
        screenshotUrls:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - category
        - categoryLabel
        - developer
        - rating
        - ratingCount
        - userCount
        - isVerified
        - isFeatured
        - tags
        - url
...
    LaunchResult:
      type: object
      properties:
        success:
          type: boolean
        url:
          type: string
      required:
        - success
        - url
...