# lib/api-zod/src/generated/types/miniApp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World App API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MiniApp {
  id: string;
  name: string;
  description: string;
  iconUrl?: string;
  category: string;
  categoryLabel: string;
  developer: string;
  rating: number;
  ratingCount: number;
  userCount: number;
  isVerified: boolean;
  isFeatured: boolean;
  tags: string[];
  url: string;
  screenshotUrls?: string[];
}